+++ Frameworks/lazyusf/lazyusf/pi/pi_controller.c
@@ -100,7 +100,7 @@ static void dma_pi_write(usf_state_t * state, struct pi_controller* pi)
     longueur = (pi->regs[PI_WR_LEN_REG] & 0xFFFFFF)+1;
     i = (pi->regs[PI_CART_ADDR_REG]-0x10000000)&0x3FFFFFF;
     longueur = (i + (int) longueur) > pi->cart_rom.rom_size ?
-               (pi->cart_rom.rom_size - i) : longueur;
+               (unsigned int)(pi->cart_rom.rom_size - i) : longueur;
     longueur = (pi->regs[PI_DRAM_ADDR_REG] + longueur) > 0x7FFFFF ?
                (0x7FFFFF - pi->regs[PI_DRAM_ADDR_REG]) : longueur;
